Delete dead code and rename ShootingStart typo (refactor batch-001)
Drops 5 unused modules with zero importers:
- src/components/InfoIcon.tsx (paired with its only consumer's only dep)
- src/utils/converter.ts (capitalizeFirst — InfoIcon was its sole caller)
- src/components/TotalProjectCard/index.tsx (superseded by PortfolioCard)
- src/sections/home/CrazyAboutSection.tsx (not referenced by HomePage)
- src/hooks/useIntersectionObserver.ts (codebase standardized on framer-motion useInView)

Renames ShootingStart.tsx -> ShootingStar.tsx (typo fix; export name was
already ShootingStar) and updates the single import in DeveloperIntro.
