--- src/utils/converter.ts
@@ -1,1 +1,0 @@
-export const capitalizeFirst = (s: string) => s ? s[0].toUpperCase() + s.slice(1) : s;
